feat: Publish types from mosaic-spec.
