+++ packages/spec/jsconfig.json
@@ -1,14 +1,10 @@
 {
+  "include": ["src/**/*", "../../specs/ts/*.ts"],
   "compilerOptions": {
-    "checkJs": false,
+    "checkJs": true,
     "noEmit": true,
     "noImplicitAny": false,
-    "module": "Node16",
-    "baseUrl": "./src",
-    "skipLibCheck": true,
-    "paths": {
-      "@uwdata/mosaic-spec": ["index"]
-    }
-  },
-  "include": ["src/**/*.d.ts", "src/**/*.js", "src/**/*.ts", "../../specs/ts/*.ts"]
+    "module": "node16",
+    "skipLibCheck": true
+  }
 }
