+++ packages/spec/src/ast-to-dom.js
@@ -1,14 +1,21 @@
+import { Param, Selection } from '@uwdata/mosaic-core';
 import { createAPIContext, loadExtension } from '@uwdata/vgplot';
+import { SpecNode } from './ast/SpecNode.js';
 import { resolveExtensions } from './config/extensions.js';
 import { error } from './util.js';
 
 /**
- * Generate running web application (DOM content) for a Mosaic spec AST.
- * @param {import('ast/SpecNode.js').SpecNode)} ast Mosaic AST root node.
+ * Generate a running web application (DOM content) for a Mosaic spec AST.
+ * @param {SpecNode} ast Mosaic AST root node.
  * @param {object} [options] Instantiation options.
  * @param {string} [options.baseURL] The base URL for loading data files.
- * @returns {Promise<import('ast-to-dom.js').AstToDOMResult>} An object with the resulting DOM element, and
- *   a map of named parameters (Param and Selection instances).
+ * @param {any[]} [options.plotDefaults] Array of default plot attributes.
+ * @param {Map<string, Param>} [options.params] A map of predefined Params/Selections.
+ * @returns {Promise<{
+ *   element: HTMLElement | SVGSVGElement;
+*    params: Map<string, Param | Selection>;
+*  }>} A Promise to an object with the resulting
+ *  DOM element, and a map of named parameters (Param and Selection instances).
  */
 export async function astToDOM(ast, options) {
   const { data, params, plotDefaults } = ast;
