+++ packages/spec/src/ast-to-esm.js
@@ -12,9 +12,9 @@ import { error, isArray, isObject, isString, toArray, toParamRef } from './util.
  *  If undefined, no connector code is generated.
  * @param {string} [options.namespace='vg'] The vgplot API namespace object.
  * @param {number} [options.depth=0] The starting indentation depth.
- * @param {Map<string,string>} [options.imports] A Map of ESM imports to
- *  include in generated code. Keys are packages (e.g., '@uwdata/vgplot')
- *  and values indicate what to import (e.g., '* as vg').
+ * @param {Map<string,string|string[]>} [options.imports] A Map of ESM
+ *  imports to include in generated code. Keys are packages (e.g.,
+ *  '@uwdata/vgplot') and values indicate what to import (e.g., '* as vg').
  * @param {string|string[]} [options.preamble] Code to include after imports.
  * @returns {string} Generated ESM code using the vgplot API.
  */
@@ -29,6 +29,7 @@ export function astToESM(ast, options = {}) {
     importsCode.push(
       isString(methods)
         ? `import ${methods} from "${pkg}";`
+        // @ts-ignore
         : `import { ${methods.join(', ')} } from "${pkg}";`
     );
   }
@@ -107,12 +108,26 @@ export function astToESM(ast, options = {}) {
 }
 
 export class CodegenContext {
+  /**
+   * Create a new code generator context.
+   * @param {object} [options] Code generation options.
+   * @param {*} [options.plotDefaults] Default attributes to apply to all plots.
+   * @param {string} [options.baseURL] The base URL for loading data files.
+   * @param {string} [options.connector] A database connector to initialize.
+   *  Valid values are 'wasm', 'socket', and 'rest'.
+   *  If undefined, no connector code is generated.
+   * @param {string} [options.namespace='vg'] The vgplot API namespace object.
+   * @param {number} [options.depth=0] The starting indentation depth.
+   * @param {Map<string,string|string[]>} [options.imports] A Map of ESM
+   *  imports to include in generated code. Keys are packages (e.g.,
+   *  '@uwdata/vgplot') and values indicate what to import (e.g., '* as vg').
+   */
   constructor({
-    plotDefaults = null,
+    plotDefaults = undefined,
     namespace = 'vg',
-    connector = null,
+    connector = undefined,
     imports = new Map([['@uwdata/vgplot', '* as vg']]),
-    baseURL = null,
+    baseURL = undefined,
     depth = 0
   } = {}) {
     this.plotDefaults = plotDefaults;
@@ -125,9 +140,11 @@ export class CodegenContext {
 
   addImport(pkg, method) {
     if (!this.imports.has(pkg)) {
-      this.imports.set(pkg, []);
+      this.imports.set(pkg, [method]);
+    } else {
+      // @ts-ignore
+      this.imports.get(pkg).push(method);
     }
-    this.imports.get(pkg).push(method);
   }
 
   setImports(pkg, all) {
