+++ packages/spec/src/ast/ASTNode.js
@@ -1,27 +1,36 @@
+/**
+ * Abstract base class for Mosaic spec AST nodes.
+ */
 export class ASTNode {
   constructor(type, children = null) {
+    /** @type {string} */
     this.type = type;
+    /** @type {ASTNode[] | null} */
     this.children = children;
   }
 
   /**
-   * @returns {void}
+   * Instantiate this AST node to use in a live web application.
+   * @param {import('../ast-to-dom.js').InstantiateContext} ctx The instantiation context.
+   * @returns {*} The instantiated value of this node.
    */
-  instantiate(/* ctx */) {
+  instantiate(ctx) { // eslint-disable-line no-unused-vars
     // @ts-ignore
     throw Error('instantiate not implemented');
   }
 
   /**
-   * @returns {string} The for for the node.
+   * Generate ESM code for this AST node.
+   * @param {import('../ast-to-esm.js').CodegenContext} ctx The code generator context.
+   * @returns {string|void} The generated ESM code for the node.
    */
-  codegen(/* ctx */) {
+  codegen(ctx) { // eslint-disable-line no-unused-vars
     // @ts-ignore
     return Error('codegen not implemented');
   }
 
   /**
-   * @returns {object} The node as JSON.
+   * @returns {*} This AST node in JSON specification format.
    */
   toJSON() {
     // @ts-ignore
