--- packages/spec/src/ast/ExpressionNode.d.ts
@@ -1,3 +1,0 @@
-export type SpecExpression =
-  | { sql: string, label?: string }
-  | { agg: string, label?: string };
