--- packages/spec/src/ast/HConcatNode.d.ts
@@ -1,8 +1,0 @@
-import { SpecComponent } from './SpecNode.js';
-
-export interface SpecHConcatNode {
-  /**
-   * Horizontally concatenate components in a row layout.
-   */
-  hconcat: SpecComponent[];
-}
