--- packages/spec/src/ast/ParamRefNode.d.ts
@@ -1,1 +1,0 @@
-export type SpecParamRef = `$${string}`;
