--- packages/spec/src/ast/SelectionNode.d.ts
@@ -1,10 +1,0 @@
-export type SelectionType =
-  | 'crossfilter'
-  | 'intersect'
-  | 'single'
-  | 'union';
-
-export interface SpecSelection {
-  select: SelectionType;
-  cross?: boolean;
-}
