--- packages/spec/src/ast/VConcatNode.d.ts
@@ -1,8 +1,0 @@
-import { SpecComponent } from './SpecNode.js';
-
-export interface SpecVConcatNode {
-  /**
-   * Vertically concatenate components in a column layout.
-   */
-  vconcat: SpecComponent[];
-}
