+++ packages/spec/src/config/inputs.js
@@ -1,5 +1,6 @@
 /**
  * Set of input widget type names.
+ * @returns {Set<string>}
  */
 export function inputNames(overrides = []) {
   return new Set([
