--- packages/spec/src/index.d.ts
@@ -1,4 +1,0 @@
-export * from './ast/SpecNode.js';
-export * from './ast-to-dom.js';
-export * from './ast-to-esm.js';
-export * from './parse-spec.js';
