+++ packages/spec/src/index.js
@@ -1,3 +1,7 @@
+/**
+ * @typedef {import('./spec/Spec.js').Spec} Spec A Mosaic declarative specification.
+ */
+
 export { astToDOM, InstantiateContext } from './ast-to-dom.js';
 export { astToESM, CodegenContext } from './ast-to-esm.js';
 export { parseSpec } from './parse-spec.js';
