+++ packages/spec/src/spec/Expression.ts
@@ -1,0 +1,13 @@
+export type Expression =
+  | SQLExpression
+  | AggregateExpression;
+
+export interface SQLExpression {
+  sql: string;
+  label?: string;
+}
+
+export interface AggregateExpression {
+  agg: string;
+  label?: string;
+}
