+++ packages/spec/src/spec/HConcat.ts
@@ -1,0 +1,8 @@
+import { Component } from './Spec.js';
+
+export interface HConcat {
+  /**
+   * Horizontally concatenate components in a row layout.
+   */
+  hconcat: Component[];
+}
