+++ packages/spec/src/spec/VConcat.ts
@@ -1,0 +1,8 @@
+import { Component } from './Spec.js';
+
+export interface VConcat {
+  /**
+   * Vertically concatenate components in a column layout.
+   */
+  vconcat: Component[];
+}
