+++ packages/spec/tsconfig.json
@@ -1,0 +1,11 @@
+{
+  "include": ["src/**/*.js", "src/**/*.ts"],
+  "compilerOptions": {
+    "allowJs": true,
+    "declaration": true,
+    "emitDeclarationOnly": true,
+    "outDir": "dist/types",
+    "module": "node16",
+    "skipLibCheck": true
+  }
+}
