+++ packages/widget/package.json
@@ -18,6 +18,7 @@
     "publish": "hatch publish"
   },
   "dependencies": {
+    "@uwdata/mosaic-spec": "0.7.1",
     "@uwdata/vgplot": "^0.7.1",
     "apache-arrow": "^15.0.0",
     "uuid": "^9.0.1"
